+++ src/Helpers/WordPress.php
@@ -28,4 +28,14 @@ class WordPress {
         return \get_the_permalink(self::getBlogArchiveID());
     }
 
+    /**
+     * Compare against WordPress version
+     * @param string $version Version to compare against.
+     * @param string $operator Operator to use.
+     */
+    public static function compareVersion($version, $operator = '>=') {
+        global $wp_version;
+        return version_compare($wp_version, $version, $operator);
+    }
+
 }