+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Remove theme support.
 * Allow to undergister patterns and categories.
 * New FA icons
+* Fix loading editor styles in Gutenberg iframe in 6.3 and newer
 
 ## 0.5.4 (2023-06-26)