+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.5.5 ()
+## 0.5.6 (2023-09-11)
+
+* Fix: no password reset emails to admin - [use correct filter](https://developer.wordpress.org/reference/functions/wp_password_change_notification/), don't disable it [for user](https://developer.wordpress.org/reference/hooks/send_password_change_email/).
+
+## 0.5.5 (2023-08-09)
 
 * Set default image size.
 * Add custom post type to ACF post type select.